Deduplicated code
